cache: Read the flags in one sweep instead of repeating it all the time
This commit introduces a new API for accessing the cached flags. Instead of
always looking up each and every flag by hand, we're now using a single sweep
through the table for all messages in a mailbox. This relies on the fact the the
items are already sorted by the UID, so the sqlite's native record stepping
works to our advantage. Unfortunately, it might well happen that the cache can
contain additional extra items (the cache on my laptop certainly does), so
there's still that fallback for accessing each item individually. This might
change in future -- perhaps the data shall be squashed to  just a single row in
the table at the cost of maybe more expensive updates.

I guess I have more measurements to make here.

Some of the tests are now a bit more noisy due to these checks; this is
something to be solved later on.

+++ src/Imap/Model/SQLCache.cpp
@@ -371,6 +371,12 @@ bool SQLCache::prepareQueries()
         return false;
     }
 
+    queryStreamedFlags = QSqlQuery(db);
+    if (!queryStreamedFlags.prepare(QLatin1String("SELECT uid, flags FROM flags WHERE mailbox = ?"))) {
+        emitError(tr("Failed to prepare queryStreamedFlags"), queryStreamedFlags);
+        return false;
+    }
+
     queryClearAllMessages1 = QSqlQuery(db);
     if (! queryClearAllMessages1.prepare(QLatin1String("DELETE FROM msg_metadata WHERE mailbox = ?"))) {
         emitError(tr("Failed to prepare queryClearAllMessages1"), queryClearAllMessages1);
@@ -694,6 +700,37 @@ void SQLCache::setMsgFlags(const QString &mailbox, const uint uid, const QString
     if (! querySetMessageFlags.exec()) {
         emitError(tr("Query querySetMessageFlags failed"), querySetMessageFlags);
     }
+}
+
+void SQLCache::prepareStreamedFlags(const QString &mailbox)
+{
+#ifdef CACHE_DEBUG
+    qDebug() << "Streamed queries for " << mailbox;
+#endif
+    queryStreamedFlags.bindValue(0, mailboxName(mailbox));
+    if (!queryStreamedFlags.exec()) {
+        emitError(tr("Query queryStreamedFlags failed"), queryStreamedFlags);
+        return;
+    }
+}
+
+StreamedUidAndFlags SQLCache::iterateStreamedFlags()
+{
+    Q_ASSERT(queryStreamedFlags.isActive());
+    if (!queryStreamedFlags.next())
+        return StreamedUidAndFlags();
+
+    uint uid = queryStreamedFlags.value(0).toUInt();
+    QStringList flags;
+    QDataStream stream(queryStreamedFlags.value(1).toByteArray());
+    stream.setVersion(streamVersion);
+    stream >> flags;
+    return StreamedUidAndFlags(uid, flags);
+}
+
+void SQLCache::freeStreamedFlags()
+{
+    queryStreamedFlags.finish();
 }
 
 AbstractCache::MessageDataBundle SQLCache::messageMetadata(const QString &mailbox, uint uid) const
